+++ CHANGELOG.md
@@ -41,6 +41,7 @@ for further information about branching and tagging conventions.
 - Removed useless fields and getters in the Redeem `Sale` and `Asset` records
 - Now using "qrcode.vue" plugin instead of "vue-qr"
 - Clipboard field styles to differ it from input fields
+- Now getting blob types from SDK
 
 #### Removed
 - "Fee" column from the movements table